iommu: qcom: iommu-v1: Properly register multiple IOMMUs
On systems having more than one IOMMU we need to register all of
them with the (DMA-)IOMMU API and to properly set their fwnode
params (and the right ops on it).

To do this, stop checking for bus_set_iommu's return value
because it is possible to register (obviously) only one
default IOMMU in the system and stop even trying to register
the IOMMU to the platform bus if there's already one: it's
gonna be fine, as long as the API has at least one registered
driver.

Also, stop checking for this driver's init to be already done:
it's not anymore possible for the kernel to try to register
the same IOMMU node twice (so that's not gonna be a problem
anymore, unlike previous kernel versions): this also allows
us to cleanly set fwnode and ops for all of this driver
instances (so, for all IOMMUs on this driver).